Update ledgermonolith tests to actually pass on build
