+++ src/ledgermonolith/src/main/java/anthos/samples/bankofanthos/ledgermonolith/BalanceReaderController.java
@@ -49,7 +49,6 @@ public final class BalanceReaderController {
     private TransactionRepository dbRepo;
 
     private String localRoutingNum;
-    private String version;
 
     private JWTVerifier verifier;
     private LoadingCache<String, Long> cache;
@@ -65,8 +64,7 @@ public final class BalanceReaderController {
             JWTVerifier verifier,
             StackdriverMeterRegistry meterRegistry,
             LoadingCache<String, Long> cache,
-            @Value("${LOCAL_ROUTING_NUM}") final String localRoutingNum,
-            @Value("${VERSION}") final String version) {
+            @Value("${LOCAL_ROUTING_NUM}") final String localRoutingNum) {
         // Initialize JWT verifier.
         this.verifier = verifier;
         LOGGER.debug("Initialized JWT verifier");
@@ -74,7 +72,6 @@ public final class BalanceReaderController {
         this.cache = cache;
         GuavaCacheMetrics.monitor(meterRegistry, this.cache, "Guava");
         LOGGER.debug("Initialized cache");
-        this.version = version;
         // Initialize transaction processor.
         this.ledgerReader = reader;
         LOGGER.debug("Initialized transaction processor");
