+++ src/ledgermonolith/src/test/java/anthos/samples/bankofanthos/ledgermonolith/LedgerMonolithControllerTest.java
@@ -1,0 +1,65 @@
+/*
+ * Copyright 2020, Google LLC.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package anthos.samples.bankofanthos.ledgermonolith;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+class LedgerMonolithControllerTest {
+
+    private LedgerMonolithController ledgerMonolithController;
+
+    private static final String VERSION = "v0.2.0";
+
+    @BeforeEach
+    void setUp() {
+        ledgerMonolithController = new LedgerMonolithController(VERSION);
+    }
+
+    @Test
+    @DisplayName("Given version number in the environment, " +
+            "return a ResponseEntity with the version number")
+    void version() {
+        // When
+        final ResponseEntity actualResult = ledgerMonolithController.version();
+
+        // Then
+        assertNotNull(actualResult);
+        assertEquals(VERSION, actualResult.getBody());
+        assertEquals(HttpStatus.OK, actualResult.getStatusCode());
+    }
+
+    @Test
+    @DisplayName("Given the server is serving requests, return HTTP Status 200")
+    void readiness() {
+        // When
+        final ResponseEntity actualResult = ledgerMonolithController.readiness();
+
+        // Then
+        assertNotNull(actualResult);
+        assertEquals(ledgerMonolithController.READINESS_CODE,
+                actualResult.getBody());
+        assertEquals(HttpStatus.OK, actualResult.getStatusCode());
+    }
+}
+
