+++ src/ledgermonolith/src/test/java/anthos/samples/bankofanthos/ledgermonolith/TransactionHistoryControllerTest.java
@@ -61,7 +61,6 @@ class TransactionHistoryControllerTest {
     @Mock
     private Deque<Transaction> transactions;
 
-    private static final String VERSION = "v0.2.0";
     private static final String LOCAL_ROUTING_NUM = "123456789";
     private static final String OK_CODE = "ok";
     private static final String JWT_ACCOUNT_KEY = "acct";
@@ -94,34 +93,10 @@ class TransactionHistoryControllerTest {
 
         when(cache.stats()).thenReturn(stats);
         transactionHistoryController = new TransactionHistoryController(ledgerReader,
-            meterRegistry, verifier, PUBLIC_KEY_PATH, cache, LOCAL_ROUTING_NUM, VERSION);
+            meterRegistry, verifier, PUBLIC_KEY_PATH, cache, LOCAL_ROUTING_NUM);
 
         when(verifier.verify(TOKEN)).thenReturn(jwt);
         when(jwt.getClaim(JWT_ACCOUNT_KEY)).thenReturn(claim);
-    }
-
-    @Test
-    @DisplayName("Given version number in the environment, " +
-            "return a ResponseEntity with the version number")
-    void version() {
-        // When
-        final ResponseEntity actualResult = transactionHistoryController.version();
-
-        // Then
-        assertNotNull(actualResult);
-        assertEquals(VERSION, actualResult.getBody());
-        assertEquals(HttpStatus.OK, actualResult.getStatusCode());
-    }
-
-    @Test
-    @DisplayName("Given the server is serving requests, return HTTP Status 200")
-    void readiness() {
-        // When
-        final String actualResult = transactionHistoryController.readiness();
-
-        // Then
-        assertNotNull(actualResult);
-        assertEquals(OK_CODE, actualResult);
     }
 
     @Test
