+++ src/ledgermonolith/src/test/java/anthos/samples/bankofanthos/ledgermonolith/TransactionValidatorTest.java
@@ -21,13 +21,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
-import static anthos.samples.bankofanthos.ledgerwriter.ExceptionMessages.
+import static anthos.samples.bankofanthos.ledgermonolith.ExceptionMessages.
         EXCEPTION_MESSAGE_INVALID_NUMBER;
-import static anthos.samples.bankofanthos.ledgerwriter.ExceptionMessages.
+import static anthos.samples.bankofanthos.ledgermonolith.ExceptionMessages.
         EXCEPTION_MESSAGE_NOT_AUTHENTICATED;
-import static anthos.samples.bankofanthos.ledgerwriter.ExceptionMessages.
+import static anthos.samples.bankofanthos.ledgermonolith.ExceptionMessages.
         EXCEPTION_MESSAGE_SEND_TO_SELF;
-import static anthos.samples.bankofanthos.ledgerwriter.ExceptionMessages.
+import static anthos.samples.bankofanthos.ledgermonolith.ExceptionMessages.
         EXCEPTION_MESSAGE_INVALID_AMOUNT;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
